requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
mlflow==2.8.1
fastapi==0.104.1
uvicorn==0.24.0
evidently==0.4.11
streamlit==1.28.2
albumentations==1.3.1
grad-cam==1.4.8
pytest==7.4.3
httpx==0.25.2
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.2